refactor mock classes into single file
