+++ tests/mutation/replacement_strategy_tests.py
@@ -1,48 +1,6 @@
-from generators.base import BaseGenerator
 from mutation.replacement_strategy import ReplaceOldest
-from pair.agent_environment_pair import Pairing
-from solvers.base import BaseSolver
+from tests.test_classes import MockGenerator, MockPair, MockSolver
 import unittest
-
-
-class MockPair(Pairing):
-    def __init__(self, solver, generator):
-        super().__init__(solver, generator)
-
-
-class MockSolver(BaseSolver):
-    def __init__(self):
-        super().__init__()
-
-    @property
-    def release(self):
-
-        class Foo:
-            def remote(self):
-                return
-
-        return Foo()
-
-
-class MockGenerator(BaseGenerator):
-    def __init__(self, expected_mutation_rate: float):
-        super().__init__()
-        self._expected_mutation_rate = expected_mutation_rate
-
-    def mutate(self, mutation_rate: float):
-        assert self._expected_mutation_rate == mutation_rate
-        return self
-
-    def update_from_lvl_string(self, level_string):
-        raise NotImplementedError
-
-    def generate_fn_wrapper(self):
-        def _generate() -> str:
-            raise NotImplementedError
-        return _generate
-
-    def __str__(self):
-        raise NotImplementedError
 
 
 class TestReplacementStrategy(unittest.TestCase):
